DEV: Mod to arith_custom docs. Removed additional operators.
Previously the docstring for arithmetic_custom included references
to additional operators such as +=, -=, *= etc. which are supposed to
simplify expressions like a=a+b, which is the equivalent of a+=b.
However, the parsing function does not evaluate these short cuts
correctly.

If these additional operators become important, then I'll revisit and
modify the parsing function, or write a custom one. However, at this
time the additional operators are deemed to be unnecessary, since the
expressions can still be evaluated without the operators.`