DEV: Changed data inputs to x1 and x2. Completed docstrings for funcs.
As per Eric's suggestion, inputs specific to data have been renamed in
order to conform to typical numpy docstrings, e.g. the arg src_data1
and src_data2 have been changed to x1 and x2.

The funcs requiring a string operation to be defined have also been
standardized so that the input order starts with the operation
or expression followed by the input data.

Finally, the docstring for arithmetic_custom was incomplete, and has
been flushed out.