DEV: Updated init.py to properly include all img proc arithmetic funcs
The tools designed to ease use of math functions in VisTrails are now
included in vttools-->to_wrap-->image_processing-->arithmetic since
these tools are explicitly designed to simplify the VisTrails user
experience. As part of the growing collection of image processing
functions included in skxray, the basic image arithmetic functions
are compiled in skxray-->image_processing-->arithmetic for
completeness.